move the build of the results-page-contents into user-votes-table-contents.util.ts
